+++ rewatch/tests/suite.sh
@@ -131,6 +131,7 @@ fi
 ./compile/17-prod-flag.sh &&
 ./compile/18-external-dep-uncurried-dot.sh &&
 ./compile/19-utf8-warning.sh &&
+./compile/20-rebuild-dependency-for-package-specs.sh &&
 ./compile/14-no-testrepo-changes.sh &&
 ./compile/15-no-new-files.sh &&
 ./compile/16-snapshots-unchanged.sh &&